update: grant-story & additional adjustments
