+++ components/Pages/Project/Review/NavbarReview.tsx
@@ -5,14 +5,12 @@ import { StarReviewIcon } from "@/components/Icons/StarReview";
 import { CardReview } from "@/components/Pages/Project/Review/CardReview";
 import { ChevronDown } from "@/components/Icons";
 import { useReviewStore } from "@/store/review";
-import {
-  getGrantStories,
-  GrantStory,
-} from "@/utilities/review/getGrantStories";
+import { getGrantStories } from "@/utilities/review/getGrantStories";
 import { useEffect } from "react";
 import { getBadge } from "@/utilities/review/getBadge";
 import { getBadgeIds } from "@/utilities/review/getBadgeIds";
 import { useSearchParams } from "next/navigation";
+import { GrantStory } from "@/types/review";
 
 export const NavbarReview = () => {
   const review = useReviewStore((state) => state.review);
@@ -20,6 +18,7 @@ export const NavbarReview = () => {
   const stories = useReviewStore((state) => state.stories);
   const setStories = useReviewStore((state) => state.setStories);
   const grantUID = useReviewStore((state) => state.grantUID);
+  const setBadge = useReviewStore((state) => state.setBadge);
   const searchParams = useSearchParams();
   const grantIdFromQueryParam = searchParams?.get("grantId");
 
@@ -30,15 +29,15 @@ export const NavbarReview = () => {
     });
   };
 
-  const setBadge = useReviewStore((state) => state.setBadge);
-
   useEffect(() => {
     const fetchGrantStories = async () => {
       if (!grantIdFromQueryParam) return;
       const grantStories = await getGrantStories(
-        grantUID ? grantUID : grantIdFromQueryParam
+        "0x635c2d0642c81e3191e6eff8623ba601b7e22e832d7791712b6bc28d052ff2b5"
+        // grantUID ? grantUID : grantIdFromQueryParam
       );
-      // setStories(grantStories); //TODO: uncomment this line to setStories
+      console.log("grantStories", grantStories);
+      setStories(grantStories);
     };
     fetchGrantStories();
   }, [grantIdFromQueryParam, grantUID]);
@@ -57,6 +56,7 @@ export const NavbarReview = () => {
     fetchData();
   }, []);
 
+  console.log("stories", stories);
   return (
     <div className="w-full flex flex-col">
       <div className="w-full flex px-2 gap-2 overflow-x-auto pb-4 relative scroller">
@@ -68,7 +68,7 @@ export const NavbarReview = () => {
               className="flex flex-col justify-center items-center text-center relative"
             >
               <p className="w-full">
-                {formatDate(new Date(storie.timestamp * 1000))}
+                {formatDate(new Date(Number(storie.timestamp) * 1000))}
               </p>
               <div className="w-full flex flex-col items-center sm:px-14 px-4">
                 <StarReviewIcon
